LPS-83478 Run Project Template Tests in parallel
