+++ modules/sdk/project-templates/project-templates/src/test/java/com/liferay/project/templates/ToolbarContributorTest.java
@@ -1,0 +1,113 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.project.templates;
+
+import java.io.File;
+
+import org.junit.Test;
+
+/**
+ * @author Lawrence Lee
+ * @author Gregory Amerson
+ * @author Andrea Di Giorgi
+ */
+public class ToolbarContributorTest extends ProjectTemplatesTest {
+
+	@Test
+	public void testBuildTemplatePortletToolbarContributor() throws Exception {
+		File gradleProjectDir = buildTemplateWithGradle(
+			"portlet-toolbar-contributor", "toolbartest", "--package-name",
+			"blade.test");
+
+		testExists(gradleProjectDir, "bnd.bnd");
+
+		testContains(
+			gradleProjectDir, "build.gradle",
+			"apply plugin: \"com.liferay.plugin\"");
+		testContains(
+			gradleProjectDir,
+			"src/main/java/blade/test/portlet/toolbar/contributor" +
+				"/ToolbartestPortletToolbarContributor.java",
+			"public class ToolbartestPortletToolbarContributor",
+			"implements PortletToolbarContributor");
+
+		File mavenProjectDir = buildTemplateWithMaven(
+			"portlet-toolbar-contributor", "toolbartest", "com.test",
+			"-DclassName=Toolbartest", "-Dpackage=blade.test");
+
+		buildProjects(gradleProjectDir, mavenProjectDir);
+	}
+
+	@Test
+	public void testBuildTemplatePortletToolbarContributor71()
+		throws Exception {
+
+		File gradleProjectDir = buildTemplateWithGradle(
+			"portlet-toolbar-contributor", "toolbartest", "--package-name",
+			"blade.test", "--liferayVersion", "7.1");
+
+		testContains(
+			gradleProjectDir, "build.gradle",
+			"name: \"com.liferay.portal.kernel\", version: \"3.0.0");
+
+		File mavenProjectDir = buildTemplateWithMaven(
+			"portlet-toolbar-contributor", "toolbartest", "com.test",
+			"-DclassName=Toolbartest", "-Dpackage=blade.test",
+			"-DliferayVersion=7.1");
+
+		testContains(
+			mavenProjectDir, "bnd.bnd", "-contract: JavaPortlet,JavaServlet");
+
+		buildProjects(gradleProjectDir, mavenProjectDir);
+	}
+
+	@Test
+	public void testBuildTemplatePortletToolbarContributorInWorkspace()
+		throws Exception {
+
+		testBuildTemplateWithWorkspace(
+			"portlet-toolbar-contributor", "blade.test",
+			"build/libs/blade.test-1.0.0.jar");
+	}
+
+	@Test
+	public void testBuildTemplatePortletToolbarContributorWithBOM()
+		throws Exception {
+
+		File gradleProjectDir = buildTemplateWithGradle(
+			"portlet-toolbar-contributor", "contributor-dependency-management",
+			"--dependency-management-enabled");
+
+		testNotContains(
+			gradleProjectDir, "build.gradle",
+			DEPENDENCY_PORTAL_KERNEL + ", version: \"2.0.0\"");
+
+		testContains(
+			gradleProjectDir, "build.gradle", DEPENDENCY_PORTAL_KERNEL);
+	}
+
+	@Test
+	public void testBuildTemplatePortletToolbarContributorWithoutBOM()
+		throws Exception {
+
+		File gradleProjectDir = buildTemplateWithGradle(
+			"portlet-toolbar-contributor", "contributor-dependency-management");
+
+		testContains(
+			gradleProjectDir, "build.gradle",
+			DEPENDENCY_PORTAL_KERNEL + ", version: \"2.0.0\"");
+	}
+
+}
